requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
ansible==6.7.0
ansible-core==2.13.13
ansible-vault==2.1.0
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
base58==2.1.1
bittensor==7.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
cytoolz==0.12.3
ddt==1.6.0
decorator==5.1.1
distro==1.9.0
ecdsa==0.19.0
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==4.4.0
eth-utils==2.2.2
fastapi==0.110.3
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
Levenshtein==0.25.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy-opentensor==0.5.0
multidict==6.0.5
munch==2.5.0
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.3
nltk==3.8.1
numpy==2.0.0
openai==1.36.0
packaging==24.1
password-strength==0.0.3.post2
pluggy==1.5.0
py==1.11.0
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
pytest==8.2.2
python-Levenshtein==0.25.1
python-statemachine==2.1.2
PyYAML==6.0.1
rapidfuzz==3.9.4
regex==2024.5.15
requests==2.32.3
resolvelib==0.8.1
retry==0.9.2
rich==13.7.1
setuptools==71.0.3
shtab==1.6.5
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
termcolor==2.4.0
toolz==0.12.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.0
websocket-client==1.8.0
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
wandb
python-dotenv
httpx[socks]